Corrección de Inicidencias de SonarQube
Se han corregido un total de 19 incidencias de SonarQube

- 3 Referidas a código duplicado (Major)
- 10 Referidas a Cobertura insuficiente de Tests (Major)
- 1 Referida a uso compacto de expresiones booleanas (Minor)
- 5 Referidas a duplicación de cadenas de texto

Refs: #13515
Refs: #13655

+++ src/main/java/py/una/pol/karaku/repo/AuditTrailDao.java
@@ -28,11 +28,11 @@ import py.una.pol.karaku.domain.AuditTrail;
 
 @Repository
 @Transactional
-public class AuditTrailDao extends KarakuBaseDao<AuditTrail, Integer> implements
-		IAuditTrailDao {
+public class AuditTrailDao extends KarakuBaseDao<AuditTrail, Long> implements
+        IAuditTrailDao {
 
-	public AuditTrailDao() {
+    public AuditTrailDao() {
 
-	}
+    }
 
 }
